[add-more-unit-tests] add FileLogger test; rename tests to say core_loggers
